+++ Assets/ViveGrip/Examples/Scripts/ViveGripExample_Hand.cs
@@ -12,8 +12,11 @@ public class ViveGripExample_Hand : MonoBehaviour {
     GetComponent<MeshFilter>().mesh = primed;
   }
 
-  void ViveGripHighlightStop() {
-    GetComponent<MeshFilter>().mesh = rest;
+  void ViveGripHighlightStop(ViveGrip_GripPoint gripPoint) {
+    // We might move out of highlight range but still be holding something
+    if (!gripPoint.HoldingSomething()) {
+      GetComponent<MeshFilter>().mesh = rest;
+    }
   }
 
   void ViveGripGrabStart() {
@@ -21,9 +24,13 @@ public class ViveGripExample_Hand : MonoBehaviour {
     StartCoroutine("FadeOut");
   }
 
-  void ViveGripGrabStop() {
+  void ViveGripGrabStop(ViveGrip_GripPoint gripPoint) {
     StopCoroutine("FadeOut");
     StartCoroutine("FadeIn");
+    // We often are touching something when we stop grabbing
+    if (!gripPoint.TouchingSomething()) {
+      GetComponent<MeshFilter>().mesh = rest;
+    }
   }
 
   IEnumerator FadeOut() {