+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+### Fixed
+- Objects will highlight when letting go of them
+- Example hand will animate properly during edge cases
 
 ## [v1.2.0] - 2016-06-11
 ### Changed